complete refactor of ellipse code - removes all duplicates, divides fns between utils, calibration, and preprocess, makes some formatting and docstring edits
