+++ py4DSTEM/process/calibration/ellipse.py
@@ -1,26 +1,8 @@
 """
-Methods related to elliptical calibration, coordinates, and transformations.
+Functions related to elliptical calibration, such as fitting elliptical
+distortions.
 
-Typically use of this module will follow a workflow like:
-(1) fit an ellipse
-(2) define an elliptical coordinate system
-(3) perform some analysis using these coordinates
-
-Methods for (1) include fit_ellipse_1d() and fit_ellipse_amorphous_ring().
-The former fits a 1D elliptical curve to a 2d arry of data e.g. a
-ring of peaks in a Bragg vector map.  The latter function fits a 2d
-pattern of ring of amorphous scattering, using a double-sided gaussian
-function intended.
-
-(2) is accomlished by feeding the output elliptical parameters of (1)
-into a Coordinates instance.
-
-(3) will depend on the dataset.  2d data can be transformed from
-carterian (x,y) space to polar-ellitical (r,phi) space.  Detected Bragg
-disk positions can be transformed from their elliptically-distorted
-cooridinates into a circularly symmetric coordinate system.
-
-The user-facing description of ellipses is in terms of the following 5
+The user-facing representation of ellipses is in terms of the following 5
 parameters:
 
     x0,y0       the center of the ellipse
@@ -30,33 +12,19 @@ parameters:
     theta       the (positive, right handed) tilt of the a-axis
                 to the x-axis, in radians
 
-Internally, fits are performed using the canonical ellipse parameterization,
-in terms of the parameters (x0,y0,A,B,C):
-
-    A(x-x0)^2 + B(x-x0)(y-y0) C(y-y0)^2 = 1
-
-It is possible to convert between (a,b,theta) <--> (A,B,C) using
-the convert_ellipse_params() methods.
-
-Transformation from cartesian to polar-elliptical space is done using
-
-       x = x0 + a*r*cos(theta)*cos(phi) + b*r*sin(theta)*sin(phi)
-       y = y0 + b*r*sin(theta)*sin(phi) - a*r*cos(theta)*cos(phi)
-
-where b = a*e is the semiminor axis length.
+More details about the elliptical parameterization used can be found in
+the module docstring for process/utils/ellipticalCoords.py.
 """
 
 import numpy as np
-from scipy.optimize import leastsq, least_squares
+from scipy.optimize import leastsq
 from scipy.ndimage.filters import gaussian_filter
-from scipy.signal import savgol_filter
-from scipy.interpolate import interp1d
 from ..utils import get_CoM
 from ...io import PointListArray
 
 ###### Fitting a 1d elliptical curve to a 2d array, e.g. a Bragg vector map ######
 
-def fit_ellipse_1d(ar,x0,y0,ri,ro,mask=None,returnABC=False):
+def fit_ellipse_1D(ar,x0,y0,ri,ro,mask=None,returnABC=False):
     """
     For a 2d array ar, fits a 1d elliptical curve to the data inside an annulus centered
     at (x0,y0) with inner and outer radii ri and ro.  The data to fit make optionally
@@ -263,276 +231,6 @@ def double_sided_gaussian(p, x, y):
     )
 
 
-### Convert between representations
-
-def convert_ellipse_params(A,B,C):
-    """
-    Converts ellipse parameters from canonical form (A,B,C) into semi-axis lengths and
-    tilt (a,e,theta), where e = b/a.
-    See module docstring for more info.
-
-    Args:
-        A,B,C (floats): parameters of an ellipse in the form:
-                             Ax^2 + Bxy + Cy^2 = 1
-
-    Returns:
-        (3-tuple): A 3-tuple consisting of:
-
-        * **a**: (float) the semimajor axis length
-        * **e**: (float) the ratio of lengths of the semiminor to the semimajor axes
-        * **theta**: (float) the tilt of the ellipse semimajor axis with respect to
-          the x-axis, in radians
-    """
-    val = np.sqrt((A-C)**2+B**2)
-    b4a = B**2 - 4*A*C
-    # Get theta
-    if B == 0:
-        if A<C:
-            theta = 0
-        else:
-            theta = np.pi/2.
-    else:
-        theta = np.arctan2((C-A-val),B)
-    # Get a,b
-    a = - np.sqrt( -2*b4a*(A+C+val) ) / b4a
-    b = - np.sqrt( -2*b4a*(A+C-val) ) / b4a
-    a,b = max(a,b),min(a,b)
-    e = b/a
-    return a,e,theta
-
-def convert_ellipse_params_r(a,e,theta):
-    """
-    Converts from ellipse parameters (a,e,theta) to (A,B,C).
-    See module docstring for more info.
-
-    Args:
-        A,B,C (floats): parameters of an ellipse in the form:
-                             Ax^2 + Bxy + Cy^2 = 1
-
-    Returns:
-        (3-tuple): A 3-tuple consisting of:
-
-        * **a**: (float) the semimajor axis length
-        * **b**: (float) the semiminor axis length
-        * **theta**: (float) the tilt of the ellipse semimajor axis with respect to
-          the x-axis, in radians
-    """
-    sin2,cos2 = np.sin(theta)**2,np.cos(theta)**2
-    b = a*e
-    a2,b2 = a**2,b**2
-    A = sin2/b2 + cos2/a2
-    C = cos2/b2 + sin2/a2
-    B = 2*(b2-a2)*np.sin(theta)*np.cos(theta)/(a2*b2)
-    return A,B,C
-
-
-
-### Polar elliptical transformations
-
-def cartesianDataAr_to_polarEllipticalDataAr(
-    cartesianData,
-    params,
-    dr=1,
-    dtheta=np.radians(2),
-    r_range=None,
-    mask=None,
-    maskThresh=0.99,
-):
-    """
-    Transforms an array of data in cartesian coordinates into a data array in
-    polar-elliptical coordinates.
-
-    If the cartesian coordinates are (qx,qy), then the parametrization for the
-    polar-elliptical coordinate system is::
-
-        qx = qx0 + A*r*cos(theta)*cos(phi) - B*r*sin(theta)*sin(phi)
-        qy = qy0 + B*r*cos(theta)*sin(phi) + A*r*sin(theta)*cos(phi)
-
-    where the final coordinates are (r,theta), and the parameters are (qx0,qy0,A,B,phi).
-    Physically, this corresponds to elliptical coordinates centered at (qx0,qy0),
-    stretched by A/B along the semimajor axes, and with those axes tilted by phi with
-    respect to the (qx,qy) axes.
-
-    Args:
-        cartesianData (2D float array): the data in cartesian coordinates
-        params (5-tuple): specifies (qx0,qy0,A,B,phi), the parameters for the
-            transformation
-        dr (float): sampling of the (r,theta) coords: the width of the bins in r
-        dtheta (float): sampling of the (r,theta) coords: the width of the bins in theta,
-            in radians
-        r_range (number or length 2 list/tuple or None): specifies the sampling of the
-            (r,theta) coords.  Precise behavior which depends on the parameter type:
-                * if r_range is a number, specifies the maximum r value
-                * if r_range is a length 2 list/tuple, specifies the min/max r values
-                * if None, autoselects max r value
-        mask (2d array of bools): shape must match cartesianData; where mask==False,
-            ignore these datapoints in making the polarElliptical data array
-        maskThresh (float): the final data mask is calculated by converting mask (above)
-            from cartesian to polar elliptical coords.  Due to interpolation, this
-            results in some non-boolean values - this is converted back to a boolean
-            array by taking polarEllipticalMask = polarTrans(mask) < maskThresh. Cells
-            where polarTrans is less than 1 (i.e. has at least one masked NN) should
-            generally be masked, hence the default value of 0.99.
-
-    Returns:
-        (3-tuple): A 3-tuple, containing:
-
-            * **polarEllipticalData**: *(2D masked array)* a masked array containing
-              the data and the data mask, in polarElliptical coordinates
-            * **rr**: *(2D array)* meshgrid of the r coordinates
-            * **tt**: *(2D array)* meshgrid of the theta coordinates
-    """
-    if mask is None:
-        mask = np.ones_like(cartesianData, dtype=bool)
-    assert (
-        cartesianData.shape == mask.shape
-    ), "Mask and cartesian data array shapes must match."
-    assert len(params) == 5, "params must have length 5"
-
-    # Get params
-    qx0, qy0, A, B, phi = params
-
-    # Define r_range: 
-    if r_range is None:
-        #find corners of image
-        corners = np.array([
-                            [0,0],
-                            [0,cartesianData.shape[0]],
-                            [0,cartesianData.shape[1]],
-                            [cartesianData.shape[0], cartesianData.shape[1]]
-                            ])
-        #find maximum corner distance
-        r_min, r_max =0, np.ceil(
-                            np.max(
-                                np.sqrt(
-                                    np.sum((corners -np.broadcast_to(np.array((qx0,qy0)), corners.shape))**2, axis = 1)
-                                       )
-                                   )
-                                ).astype(int)
-    else:
-        try:
-            r_min, r_max = r_range[0], r_range[1]
-        except TypeError:
-            r_min, r_max = 0, r_range
-
-    Nx, Ny = cartesianData.shape
-    # Define the r/theta coords
-    r_bins = np.arange(r_min + dr / 2.0, r_max + dr / 2.0, dr)  # values are bin centers
-    t_bins = np.arange(-np.pi + dtheta / 2.0, np.pi + dtheta / 2.0, dtheta)
-    rr, tt = np.meshgrid(r_bins, t_bins)
-    Nr, Nt = rr.shape
-
-    # Get (qx,qy) corresponding to each (r,theta) in the newly defined coords
-    xr = rr * np.cos(tt)
-    yr = rr * np.sin(tt)
-    qx = qx0 + xr * A * np.cos(phi) - yr * B * np.sin(phi)
-    qy = qy0 + xr * A * np.sin(phi) + yr * B * np.cos(phi)
-
-    # qx,qy are now shape (Nr,Ntheta) arrays, such that (qx[r,theta],qy[r,theta]) is the point
-    # in cartesian space corresponding to r,theta.  We now get the values for the final
-    # polarEllipticalData array by interpolating values at these coords from the original
-    # cartesianData array.
-
-    transform_mask = (qx > 0) * (qy > 0) * (qx < Nx - 1) * (qy < Ny - 1)
-
-    # Bilinear interpolation
-    xF = np.floor(qx[transform_mask])
-    yF = np.floor(qy[transform_mask])
-    dx = qx[transform_mask] - xF
-    dy = qy[transform_mask] - yF
-    x_inds = np.vstack((xF, xF + 1, xF, xF + 1)).astype(int)
-    y_inds = np.vstack((yF, yF, yF + 1, yF + 1)).astype(int)
-    weights = np.vstack(
-        ((1 - dx) * (1 - dy), (dx) * (1 - dy), (1 - dx) * (dy), (dx) * (dy))
-    )
-    transform_mask = transform_mask.ravel()
-    polarEllipticalData = np.zeros(Nr * Nt)
-    polarEllipticalData[transform_mask] = np.sum(
-        cartesianData[x_inds, y_inds] * weights, axis=0
-    )
-    polarEllipticalData = np.reshape(polarEllipticalData, (Nr, Nt))
-
-    # Transform mask
-    polarEllipticalMask = np.zeros(Nr * Nt)
-    polarEllipticalMask[transform_mask] = np.sum(mask[x_inds, y_inds] * weights, axis=0)
-    polarEllipticalMask = np.reshape(polarEllipticalMask, (Nr, Nt))
-
-    polarEllipticalData = np.ma.array(
-        data=polarEllipticalData, mask=polarEllipticalMask < maskThresh
-    )
-    return polarEllipticalData, rr, tt
-
-
-### Radial integration
-
-def radial_integral(ar, x0, y0, dr):
-    """
-    Computes the radial integral of array ar from center (x0,y0) with a step size in r of dr.
-
-    Args:
-        ar (2d array): the data
-        x0,y0 (floats): the origin
-
-    Returns:
-        (2-tuple): A 2-tuple containing:
-
-            * **rbin_centers**: *(1d array)* the bins centers of the radial integral
-            * **radial_integral**: *(1d array)* the radial integral
-    """
-    rmax = int(
-        max(
-            (
-                np.hypot(x0, y0),
-                np.hypot(x0, ar.shape[1] - y0),
-                np.hypot(ar.shape[0] - x0, y0),
-                np.hypot(ar.shape[0] - x0, ar.shape[1] - y0),
-            )
-        )
-    )
-    polarAr, rr, tt = cartesianDataAr_to_polarEllipticalDataAr(
-        ar, params=(x0, y0, 1, 1, 0), dr=dr, dtheta=np.radians(2), r_range=rmax
-    )
-    radial_integral = np.sum(polarAr, axis=0)
-    rbin_centers = rr[0, :]
-    return rbin_centers,radial_integral
-
-
-def radial_elliptical_integral(ar, dr, ellipse_params):
-    """
-    Computes the radial integral of array ar from center (x0,y0) with a step size in r of dr.
-
-    Args:
-        ar (2d array): the data
-        dr (number): the r sampling
-        ellipse_params (5-tuple): the parameters (x0,y0,A,B,phi) for the ellipse
-
-    Returns:
-        (2-tuple): A 2-tuple containing:
-
-            * **rbin_centers**: *(1d array)* the bins centers of the radial integral
-            * **radial_integral**: *(1d array)* the radial integral
-        radial_integral (1d array) the radial integral
-    """
-    x0, y0 = ellipse_params[0], ellipse_params[1]
-    rmax = int(
-        max(
-            (
-                np.hypot(x0, y0),
-                np.hypot(x0, ar.shape[1] - y0),
-                np.hypot(ar.shape[0] - x0, y0),
-                np.hypot(ar.shape[0] - x0, ar.shape[1] - y0),
-            )
-        )
-    )
-    polarAr, rr, tt = cartesianDataAr_to_polarEllipticalDataAr(
-        ar, params=ellipse_params, dr=dr, dtheta=np.radians(2), r_range=rmax
-    )
-    radial_integral = np.sum(polarAr, axis=0)
-    rbin_centers = rr[0, :]
-    return rbin_centers,radial_integral
-
-
-
 ### Correct Bragg peak positions, making a circular coordinate system
 
 def correct_braggpeak_elliptical_distortions(braggpeaks,e,theta,x0=0,y0=0):
@@ -572,7 +270,6 @@ def correct_braggpeak_elliptical_distortions(braggpeaks,e,theta,x0=0,y0=0):
             pointlist.data["qx"] = xyar_f[0, :] + x0
             pointlist.data["qy"] = xyar_f[1, :] + y0
     return braggpeaks_corrected
-
 
 
 ### Fit an ellipse to crystalline scattering with a known angle between peaks
@@ -642,137 +339,5 @@ def constrain_degenerate_ellipse(data, x, y, a, b, theta, r_inner, r_outer, phi_
     return a_constrained, b_constrained
 
 
-## MOVE TO VIS
-def compare_double_sided_gaussian(data, p, power=1, mask=None):
-    """
-    Plots a comparison between a diffraction pattern and a fit, given p.
-    """
-    if mask is None:
-        mask = np.ones_like(data)
 
-    yy, xx = np.meshgrid(np.arange(data.shape[1]), np.arange(data.shape[0]))
-    data_fit = double_sided_gaussian(p, xx, yy)
-
-    theta = np.arctan2(xx - p[7], yy - p[8])
-    theta_mask = np.cos(theta * 8) > 0
-    data_combined = (data * theta_mask + data_fit * (1 - theta_mask)) ** power
-    data_combined = mask * data_combined
-    plt.figure(12, clear=True)
-    plt.imshow(data_combined)
-
-    return
-
-## Create look up table for background subtraction
-def get_1Dbackground(data, qx0, qy0 ,e, phi,
-                    maskUpdateIter=3,
-                    min_relative_threshold = 4,
-                    smoothing = False,
-                    smoothingWindowSize = 3,
-                    smoothingPolyOrder = 4,
-                    smoothing_log = True,
-                    min_background_value=1E-3,
-                    return_polararr=False):
-    """
-    Gets the median polar background for a diffraction pattern
-
-    Accepts:
-    data                    (ndarray) the data for which to find the polar eliptical background, usually a diffraction pattern
-    qx0,qy0                 (floats) the ellipse center; if the braggpeaks have been centered,
-                            these should be zero
-    e                       (float) the length ratio of semiminor/semimajor axes
-    phi                     (float) tilt of the major axis with respect  to (qx, qy) axis, in radians
-    maskUpdate_iter         (integer)
-    min_relative_threshold  (float)
-    smoothing               (bool) if true savgol filter smoothing is applied
-    smoothingWindowSize     (integer) size of the smoothing window, must be odd number
-    smoothingPolyOrder      (number) order of the polynomial smoothing to be applied
-    smoothing_log           (bool) if true log smoothing is performed
-    min_background_value    (float) if log smoothing is true, a zero value will be replaced with a small nonzero float
-    return_polar_arr        (bool) if True the polar transform with the masked high intensity peaks will be returned
-
-    Returns:
-    background1D (ndarray) 1D polar elliptical background
-    r_bins       (ndarray) the elliptically transformed radius associated with background1D
-    polarData    (ndarray) **optional: returns the masked polar transform from which the 1D background is computed
-    """
-    # assert data is proper form 
-    assert isinstance(smoothing, bool), "Smoothing must be bool"
-    assert smoothingWindowSize%2==1, 'Smoothing window must be odd'
-    assert isinstance(return_polararr, bool), "return_polararr must be bool"
-
-    # Compute Polar Transform
-    polarData, rr, tt = cartesianDataAr_to_polarEllipticalDataAr(data,tuple([qx0,qy0, 1, e, phi]))
-
-    # Crop polar data to maximum distance which contains information from original image
-    if (polarData.mask.sum(axis = (0))==polarData.shape[0]).any():
-            ii = polar.data.shape[1]
-            while(polar.mask[:,ii].all()==True):
-                ii = ii-1
-            maximalDistance = ii
-            polarData = polarData[:,0:maximalDistance]
-            r_bins = rr[0,0:maximalDistance]
-    else:
-            r_bins = rr[0,:]
-
-    # Iteratively mask off high intensity peaks
-    maskPolar = np.copy(polarData.mask)
-    for ii in range(maskUpdateIter+1):
-        if ii > 0:
-            maskUpdate = np.logical_or(maskPolar,
-                                        polarData/background1D > min_relative_threshold)
-            # Prevent entire columns from being masked off 
-            colMaskMin = np.all(maskUpdate, axis = 0)  # Detect columns that are empty
-            maskUpdate[:,colMaskMin] = polarData.mask[:,colMaskMin] # reset empty columns to values of previous iterations
-            polarData.mask  = maskUpdate  # Update Mask
-
-        background1D = np.ma.median(polarData, axis = 0)
-
-    background1D = np.maximum(background1D, min_background_value)
-
-    if smoothing == True:
-        if smoothing_log==True:
-            background1D = np.log(background1D)
-
-        background1D = savgol_filter(background1D,
-                                     smoothingWindowSize,
-                                     smoothingPolyOrder)
-        if smoothing_log==True:
-            background1D = np.exp(background1D)
-    if return_polararr ==True:
-        return(background1D, r_bins, polarData)
-    else:
-        return(background1D, r_bins)
-
-#Create 2D Background 
-def get_2Dbackground(data, background1D, r_bins, qx0, qy0, phi, e):
-    """
-    Gets 2D polar elliptical background from linear 1D background
-
-    Accepts:
-    data            (ndarray) the data for which to find the polar eliptical background, usually a diffraction pattern
-    background1D:   (ndarray) a vector representing the radial elliptical background
-    r_bins:         (ndarray) a vector of the elliptically transformed radius associated with background1D
-    qx0,qy0         (floats) the ellipse center; if the braggpeaks have been centered,
-                              these should be zero
-    e               (float) the length ratio of semiminor/semimajor axes
-    phi             (float) tilt of the major axis with respect  to (qx, qy) axis, in radians
-
-    Returns:
-    2D Background   (ndarray) 2D polar elliptical median background image
-    """
-    assert r_bins.shape==background1D.shape, "1D background and r_bins must be same length"
-    # Define centered 2D cartesian coordinate system
-    yc, xc = np.meshgrid(np.arange(0,data.shape[1])-qy0,
-                         np.arange(0,data.shape[0])-qx0)
-
-
-    # Calculate the semimajor axis distance for each point in the 2D array
-    r = np.sqrt(((xc*np.cos(phi)+yc*np.sin(phi))**2)+
-                (((xc*np.sin(phi)-yc*np.cos(phi))**2)/(e**2)))
-
-    # Create a 2D eliptical background using linear interpolation  
-    f = interp1d(r_bins, background1D, fill_value = 'extrapolate')
-    background2D = f(r)
-
-    return(background2D)
